requirements.txt
fastapi==0.111.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
streamlit==1.36.0
torch
transformers 
accelerate 
bitsandbytes
weaviate
colpali-engine
pymupdf 
pdf2image
qdrant-client
vidore-benchmark
deepeval